pyproject.toml
[project]
name = "iqr_vision"
version = "0.0.1"
authors = [
  { name="Neil Song", email="neil.song@yale.edu" },
]
description = "A collection of vision utilities for IQR Lab"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
dependencies = [
  "python-dotenv>=1.0.1",
  "typer>=0.15.1",
  "rosbags>=0.10.7",
  "pillow>=11.1.0",
  "numpy>=2.2.2",
  "tqdm>=4.67.1"
]

[project.urls]
Homepage = "https://github.com/iqr-lab/vision-utils/"
Issues = "https://github.com/iqr-lab/vision-utils/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
iqr-vision = "iqr_vision.cli:app"